structureTreeRecord route: no special case for structures.
The old code first checked whether the target content item was a
structure, and if so, made sure its ancestory tree matched the current
route. This, however, produced wrong results with nested structure
trees, and broke the breadcrumbs function.

Removed the special case check for structures, just accepting the slug
if a matching record exists, and ignoring the path segments before it.
This causes other URLs to also match, even when the path is technically
incorrect (e.g., /beeble/brox/something-real will reference
"something-real" even if no structure items slugged "beeble" or "brox"
exist, and also if those exist but do not match the actual ancestory
tree of the "something-real" record). For any contenttype other than
structures, this behavior was already in place, and seeing how
structures themselves are also just a contenttype, I do not see a good
reason to treat them specially in this context.